// livekit-rtc/src/proto/video_frame_pb.ts
export enum VideoRotation {
  /**
   * @generated from enum value: VIDEO_ROTATION_0 = 0;
   */
  VIDEO_ROTATION_0 = 0,

  /**
   * @generated from enum value: VIDEO_ROTATION_90 = 1;
   */
  VIDEO_ROTATION_90 = 1,

  /**
   * @generated from enum value: VIDEO_ROTATION_180 = 2;
   */
  VIDEO_ROTATION_180 = 2,

  /**
   * @generated from enum value: VIDEO_ROTATION_270 = 3;
   */
  VIDEO_ROTATION_270 = 3,
}
// Retrieve enum metadata with: proto3.getEnumType(VideoRotation)
proto3.util.setEnumType(VideoRotation, "livekit.proto.VideoRotation", [
  { no: 0, name: "VIDEO_ROTATION_0" },
  { no: 1, name: "VIDEO_ROTATION_90" },
  { no: 2, name: "VIDEO_ROTATION_180" },
  { no: 3, name: "VIDEO_ROTATION_270" },
]);

/**
 * @generated from enum livekit.proto.VideoFormatType
 */
export enum VideoFormatType {
  /**
   * @generated from enum value: FORMAT_ARGB = 0;
   */
  FORMAT_ARGB = 0,

  /**
   * @generated from enum value: FORMAT_BGRA = 1;
   */
  FORMAT_BGRA = 1,

  /**
   * @generated from enum value: FORMAT_ABGR = 2;
   */
  FORMAT_ABGR = 2,

  /**
   * @generated from enum value: FORMAT_RGBA = 3;
   */
  FORMAT_RGBA = 3,
}
// Retrieve enum metadata with: proto3.getEnumType(VideoFormatType)
proto3.util.setEnumType(VideoFormatType, "livekit.proto.VideoFormatType", [
  { no: 0, name: "FORMAT_ARGB" },
  { no: 1, name: "FORMAT_BGRA" },
  { no: 2, name: "FORMAT_ABGR" },
  { no: 3, name: "FORMAT_RGBA" },
]);

/**
 * @generated from enum livekit.proto.VideoFrameBufferType
 */
export enum VideoFrameBufferType {
  /**
   * @generated from enum value: NATIVE = 0;
   */
  NATIVE = 0,

  /**
   * @generated from enum value: I420 = 1;
   */
  I420 = 1,

  /**
   * @generated from enum value: I420A = 2;
   */
  I420A = 2,

  /**
   * @generated from enum value: I422 = 3;
   */
  I422 = 3,

  /**
   * @generated from enum value: I444 = 4;
   */
  I444 = 4,

  /**
   * @generated from enum value: I010 = 5;
   */
  I010 = 5,

  /**
   * @generated from enum value: NV12 = 6;
   */
  NV12 = 6,
}
// ...
export class ToI420Request extends Message<ToI420Request> {
  /**
   * @generated from field: bool flip_y = 1;
   */
  flipY = false;

  /**
   * @generated from oneof livekit.proto.ToI420Request.from
   */
  from: {
    /**
     * @generated from field: livekit.proto.ArgbBufferInfo argb = 2;
     */
    value: ArgbBufferInfo;
    case: "argb";
  } | {
    /**
     * @generated from field: livekit.proto.VideoFrameBufferInfo buffer = 3;
     */
    value: VideoFrameBufferInfo;
    case: "buffer";
  } | {
    /**
     * @generated from field: uint64 handle = 4;
     */
    value: bigint;
    case: "handle";
  } | { case: undefined; value?: undefined } = { case: undefined };

  constructor(data?: PartialMessage<ToI420Request>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "livekit.proto.ToI420Request";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "flip_y", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
    { no: 2, name: "argb", kind: "message", T: ArgbBufferInfo, oneof: "from" },
    { no: 3, name: "buffer", kind: "message", T: VideoFrameBufferInfo, oneof: "from" },
    { no: 4, name: "handle", kind: "scalar", T: 4 /* ScalarType.UINT64 */, oneof: "from" },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): ToI420Request {
    return new ToI420Request().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): ToI420Request {
    return new ToI420Request().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): ToI420Request {
    return new ToI420Request().fromJsonString(jsonString, options);
  }

  static equals(a: ToI420Request | PlainMessage<ToI420Request> | undefined, b: ToI420Request | PlainMessage<ToI420Request> | undefined): boolean {
    return proto3.util.equals(ToI420Request, a, b);
  }
}

/**
 * @generated from message livekit.proto.ToI420Response
 */
export class ToI420Response extends Message<ToI420Response> {
  /**
   * @generated from field: livekit.proto.OwnedVideoFrameBuffer buffer = 1;
   */
  buffer?: OwnedVideoFrameBuffer;

  constructor(data?: PartialMessage<ToI420Response>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "livekit.proto.ToI420Response";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "buffer", kind: "message", T: OwnedVideoFrameBuffer },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): ToI420Response {
    return new ToI420Response().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): ToI420Response {
    return new ToI420Response().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): ToI420Response {
    return new ToI420Response().fromJsonString(jsonString, options);
  }

  static equals(a: ToI420Response | PlainMessage<ToI420Response> | undefined, b: ToI420Response | PlainMessage<ToI420Response> | undefined): boolean {
    return proto3.util.equals(ToI420Response, a, b);
  }
}
// ...
export class OwnedVideoFrameBuffer extends Message<OwnedVideoFrameBuffer> {
  /**
   * @generated from field: livekit.proto.FfiOwnedHandle handle = 1;
   */
  handle?: FfiOwnedHandle;

  /**
   * @generated from field: livekit.proto.VideoFrameBufferInfo info = 2;
   */
  info?: VideoFrameBufferInfo;

  constructor(data?: PartialMessage<OwnedVideoFrameBuffer>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "livekit.proto.OwnedVideoFrameBuffer";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "handle", kind: "message", T: FfiOwnedHandle },
    { no: 2, name: "info", kind: "message", T: VideoFrameBufferInfo },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): OwnedVideoFrameBuffer {
    return new OwnedVideoFrameBuffer().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): OwnedVideoFrameBuffer {
    return new OwnedVideoFrameBuffer().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): OwnedVideoFrameBuffer {
    return new OwnedVideoFrameBuffer().fromJsonString(jsonString, options);
  }

  static equals(a: OwnedVideoFrameBuffer | PlainMessage<OwnedVideoFrameBuffer> | undefined, b: OwnedVideoFrameBuffer | PlainMessage<OwnedVideoFrameBuffer> | undefined): boolean {
    return proto3.util.equals(OwnedVideoFrameBuffer, a, b);
  }
}

/**
 * @generated from message livekit.proto.PlanarYuvBufferInfo
 */
export class PlanarYuvBufferInfo extends Message<PlanarYuvBufferInfo> {
  /**
   * @generated from field: uint32 chroma_width = 1;
   */
  chromaWidth = 0;

  /**
   * @generated from field: uint32 chroma_height = 2;
   */
  chromaHeight = 0;

  /**
   * @generated from field: uint32 stride_y = 3;
   */
  strideY = 0;

  /**
   * @generated from field: uint32 stride_u = 4;
   */
  strideU = 0;

  /**
   * @generated from field: uint32 stride_v = 5;
   */
  strideV = 0;

  /**
   * @generated from field: uint32 stride_a = 6;
   */
  strideA = 0;

  /**
   * *const u8 or *const u16
   *
   * @generated from field: uint64 data_y_ptr = 7;
   */
  dataYPtr = protoInt64.zero;

  /**
   * @generated from field: uint64 data_u_ptr = 8;
   */
  dataUPtr = protoInt64.zero;

  /**
   * @generated from field: uint64 data_v_ptr = 9;
   */
  dataVPtr = protoInt64.zero;

  /**
   * nullptr = no alpha
   *
   * @generated from field: uint64 data_a_ptr = 10;
   */
  dataAPtr = protoInt64.zero;

  constructor(data?: PartialMessage<PlanarYuvBufferInfo>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "livekit.proto.PlanarYuvBufferInfo";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "chroma_width", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 2, name: "chroma_height", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 3, name: "stride_y", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 4, name: "stride_u", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 5, name: "stride_v", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 6, name: "stride_a", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 7, name: "data_y_ptr", kind: "scalar", T: 4 /* ScalarType.UINT64 */ },
    { no: 8, name: "data_u_ptr", kind: "scalar", T: 4 /* ScalarType.UINT64 */ },
    { no: 9, name: "data_v_ptr", kind: "scalar", T: 4 /* ScalarType.UINT64 */ },
    { no: 10, name: "data_a_ptr", kind: "scalar", T: 4 /* ScalarType.UINT64 */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): PlanarYuvBufferInfo {
    return new PlanarYuvBufferInfo().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): PlanarYuvBufferInfo {
    return new PlanarYuvBufferInfo().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): PlanarYuvBufferInfo {
    return new PlanarYuvBufferInfo().fromJsonString(jsonString, options);
  }

  static equals(a: PlanarYuvBufferInfo | PlainMessage<PlanarYuvBufferInfo> | undefined, b: PlanarYuvBufferInfo | PlainMessage<PlanarYuvBufferInfo> | undefined): boolean {
    return proto3.util.equals(PlanarYuvBufferInfo, a, b);
  }
}

/**
 * @generated from message livekit.proto.BiplanarYuvBufferInfo
 */
export class BiplanarYuvBufferInfo extends Message<BiplanarYuvBufferInfo> {
  /**
   * @generated from field: uint32 chroma_width = 1;
   */
  chromaWidth = 0;

  /**
   * @generated from field: uint32 chroma_height = 2;
   */
  chromaHeight = 0;

  /**
   * @generated from field: uint32 stride_y = 3;
   */
  strideY = 0;

  /**
   * @generated from field: uint32 stride_uv = 4;
   */
  strideUv = 0;

  /**
   * @generated from field: uint64 data_y_ptr = 5;
   */
  dataYPtr = protoInt64.zero;

  /**
   * @generated from field: uint64 data_uv_ptr = 6;
   */
  dataUvPtr = protoInt64.zero;

  constructor(data?: PartialMessage<BiplanarYuvBufferInfo>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "livekit.proto.BiplanarYuvBufferInfo";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "chroma_width", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 2, name: "chroma_height", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 3, name: "stride_y", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 4, name: "stride_uv", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 5, name: "data_y_ptr", kind: "scalar", T: 4 /* ScalarType.UINT64 */ },
    { no: 6, name: "data_uv_ptr", kind: "scalar", T: 4 /* ScalarType.UINT64 */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): BiplanarYuvBufferInfo {
    return new BiplanarYuvBufferInfo().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): BiplanarYuvBufferInfo {
    return new BiplanarYuvBufferInfo().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): BiplanarYuvBufferInfo {
    return new BiplanarYuvBufferInfo().fromJsonString(jsonString, options);
  }

  static equals(a: BiplanarYuvBufferInfo | PlainMessage<BiplanarYuvBufferInfo> | undefined, b: BiplanarYuvBufferInfo | PlainMessage<BiplanarYuvBufferInfo> | undefined): boolean {
    return proto3.util.equals(BiplanarYuvBufferInfo, a, b);
  }
}